+++ src/base/promptRenderer.ts
@@ -580,7 +580,13 @@ export class PromptRenderer<P extends BasePromptElementProps> {
 	}
 
 	private _handleIntrinsicElementJSON(node: PromptTreeElement, data: JSONT.PromptElementJSON) {
-		node.appendPieceJSON(data.node);
+		const appended = node.appendPieceJSON(data.node);
+		if (this.tracer?.includeInEpoch) {
+			for (const child of appended.elements()) {
+				// tokenBudget is just 0 because we don't know the renderer state on the tool side.
+				this.tracer.includeInEpoch({ id: child.id, tokenBudget: 0 });
+			}
+		}
 	}
 
 	private _handleIntrinsicUsedContext(
@@ -911,7 +917,7 @@ class PromptTreeElement {
 				this._obj.props.name,
 				this._obj instanceof AssistantMessage ? this._obj.props.toolCalls : undefined,
 				this._obj instanceof ToolMessage ? this._obj.props.toolCallId : undefined,
-				this._obj.props.priority ?? 0,
+				this._obj.props.priority ?? Number.MAX_SAFE_INTEGER,
 				this._metadata,
 				this._children.map(child => child.materialize())
 			);
@@ -924,7 +930,7 @@ class PromptTreeElement {
 			return new MaterializedContainer(
 				this.id,
 				this._obj?.constructor.name,
-				this._obj?.props.priority || 0,
+				this._obj?.props.priority ?? Number.MAX_SAFE_INTEGER,
 				this._children.map(child => child.materialize()),
 				this._metadata,
 				flags
@@ -934,6 +940,15 @@ class PromptTreeElement {
 
 	public addMetadata(metadata: PromptMetadata): void {
 		this._metadata.push(metadata);
+	}
+
+	public *elements(): Iterable<PromptTreeElement> {
+		yield this;
+		for (const child of this._children) {
+			if (child instanceof PromptTreeElement) {
+				yield* child.elements();
+			}
+		}
 	}
 }